Remove the dependency on Hashie::Rash gem
This gem is no longer maintained and is a blocker to upgrading to the
latest version of the Hashie gem as there is now a namespace conflict
with a new module Rash that Hashie also includes.

This simply adds the functionality to convert from camelcase to
underscores as part of the Faraday request in to the gem itself (rather
than relying on external dependencies).